pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "disaster-fact-graph"
version = "0.1.0"
description = "FEMA Individual Assistance eligibility rules as an executable Fact Graph"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "mypy>=1.10.0",
    "pydantic[mypy]>=2.0.0",
]

[project.scripts]
fact-graph = "cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["fact_graph*", "pipeline*", "cli*"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
warn_return_any = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "synthetic_cases.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
